memberNo int -> long 수정

+++ finalproject6/src/main/java/com/kh/final6/entity/MemberDto.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 
 @Data @NoArgsConstructor @AllArgsConstructor @Builder
 public class MemberDto {
-	private int memberNo;
+	private long memberNo;
 	private String memberId;
 	private String memberPw;
 	private String memberName;
